renamed installer package to config for better naming, moved NewClient funcs to wrapped.go
