+++ development/prow-installer/config/prow-installer-config.yaml
@@ -7,12 +7,9 @@ bucket_name: test-prow-bucket
 keyring_name: test-prow-keyring
 encryption_key_name: test-prow-key
 kubeconfig: ""
-prefix: kamil
+prefix: ""
 labels:
   created-by: user-name
-  random-test-label: helloworld
-#  created-on: xxx
-#  ttl: xxx
 serviceAccounts:
   - name: sa-gke-kyma-integration # present
     type: serviceAccount
